requirements.txt
acdh-django-browsing>=2.0,<3
acdh-django-charts>=0.5.3,<1
django-autocomplete-light>=3.8.1
crispy-bootstrap5
django-leaflet>=0.30.1,<1
django-reversion>=4.0.2,<4.1
Django>=5.0,<6
djangorestframework-gis
geojson
geopandas
requests>=2.25
psycopg2-binary
coreapi>=2.3.3
rdflib>=7.0.0,<8